backend/requirements.txt

fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy>=2.0.23
psycopg2-binary
python-dotenv==1.0.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
pyjwt>=2.8.0
python-multipart==0.0.6
parsel>=1.8.1
pytest==7.4.3
httpx==0.25.2

